Reduce HandiNAVI to a resource dependency point with a nice interface.
Drastically simplified the application configuration by both
bundle-izing the phantasm at rest services, but creating a common
dropwizard application at that level that can be reused for tests and of
course HandiNAVI.
