+++ navi/pom.xml
@@ -70,10 +70,6 @@
             <groupId>com.chiralbehaviors.CoRE</groupId>
             <artifactId>browser</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.chiralbehaviors.CoRE</groupId>
-            <artifactId>loader</artifactId>
-        </dependency>
     </dependencies>
     <build>
         <plugins>
