+++ phantasm-at-rest/src/main/java/com/chiralbehaviors/CoRE/phantasm/service/JpaConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License.
  */
-package com.chiralbehaviors.CoRE.navi;
+package com.chiralbehaviors.CoRE.phantasm.service;
 
 import java.util.HashMap;
 import java.util.Map;
