--- phantasm-at-rest/src/test/java/com/chiralbehaviors/CoRE/phantasm/resources/test/JpaConfiguration.java
@@ -1,53 +1,0 @@
-/** 
- * (C) Copyright 2012 Chiral Behaviors, LLC. All Rights Reserved
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- *     
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, 
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
- * See the License for the specific language governing permissions and 
- * limitations under the License.
- */
-package com.chiralbehaviors.CoRE.phantasm.resources.test;
-
-import io.dropwizard.Configuration;
-
-import java.util.Map;
-
-import org.hibernate.validator.constraints.NotEmpty;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * @author hhildebrand
- * 
- */
-public class JpaConfiguration extends Configuration {
-    @JsonProperty
-    private boolean debug = false;
-
-    @NotEmpty
-    @JsonProperty
-    private String persistenceUnit;
-
-    @NotEmpty
-    @JsonProperty
-    private Map<String, String> properties;
-
-    public String getPersistenceUnit() {
-        return persistenceUnit;
-    }
-
-    public Map<String, String> getProperties() {
-        return properties;
-    }
-
-    public boolean isDebug() {
-        return debug;
-    }
-}
