+++ phantasm-at-rest/src/test/resources/demo.yml
@@ -1,16 +1,11 @@
+auth: NULL
+
 jpa:
-    persistenceUnit: CoRE
     properties: 
         javax.persistence.jdbc.user: ${db.login}
         javax.persistence.jdbc.password: ${db.password}
         javax.persistence.jdbc.url: jdbc:postgresql://${db.server}:${db.port}/${db.database}
         javax.persistence.jdbc.driver: org.postgresql.Driver
-        hibernate.c3p0.max_size: 20
-        hibernate.c3p0.min_size: 10
-        hibernate.c3p0.timeout: 5000
-        hibernate.c3p0.max_statements: 100
-        hibernate.c3p0.acquire_increment: 2
-        hibernate.dialect: com.chiralbehaviors.CoRE.attribute.json.JsonPostgreSqlDialect
         
 logging:
     level: INFO
