+++ phantasm-at-rest/src/test/resources/oauth.yml
@@ -1,6 +1,6 @@
 randomPort: true 
 
-useOauth: true
+auth: BEARER_TOKEN
 
 logging:
     level: ERROR
