+++ phantasm-at-rest/src/test/resources/test.yml
@@ -1,3 +1,5 @@
+auth: NULL
+
 randomPort: true 
         
 logging:
